+++ lib/custom_widgets/feedback_photo_card_widget.dart
@@ -12,6 +12,10 @@ class FeedbackPhotoCardWidget extends StatelessWidget {
   final double cardHeight;
   final double cardWidth;
   final bool hideCenterButton;
+  final bool hideTitleText;
+  final bool hideDescriptionText;
+  final BoxFit imageScaleType;
+  final Color? imageBackgroundColor;
   final FeedbackPhotoCardValueNotifier feedbackPhotoCardValueNotifier;
   final IconData? leftButtonIcon;
   final IconData? centerButtonIcon;
@@ -28,6 +32,10 @@ class FeedbackPhotoCardWidget extends StatelessWidget {
     required this.cardHeight,
     required this.cardWidth,
     required this.hideCenterButton,
+    required this.hideTitleText,
+    required this.hideDescriptionText,
+    required this.imageScaleType,
+    required this.imageBackgroundColor,
     required this.feedbackPhotoCardValueNotifier,
     this.leftButtonIcon,
     this.centerButtonIcon,
@@ -67,18 +75,19 @@ class FeedbackPhotoCardWidget extends StatelessWidget {
                   clipBehavior: Clip.hardEdge,
                   decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(25.0),
+                    color: imageBackgroundColor,
                     image: photoCard.isLocalImage
                         ? DecorationImage(
                             image: AssetImage(
                               photoCard.imagePath,
                             ),
-                            fit: BoxFit.cover,
+                            fit: imageScaleType,
                           )
                         : DecorationImage(
                             image: NetworkImage(
                               photoCard.imagePath,
                             ),
-                            fit: BoxFit.cover,
+                            fit: imageScaleType,
                           ),
                   ),
                   margin: const EdgeInsets.all(10.0),
@@ -159,11 +168,17 @@ class FeedbackPhotoCardWidget extends StatelessWidget {
                   SizedBox(
                     height: 20.0,
                   ),
-                  PhotoTitleWidget(photoCard: photoCard),
+                  Visibility(
+                    visible: !hideTitleText,
+                    child: PhotoTitleWidget(photoCard: photoCard),
+                  ),
                   SizedBox(
                     height: 5.0,
                   ),
-                  PhotoDescriptionWidget(photoCard: photoCard),
+                  Visibility(
+                    visible: !hideDescriptionText,
+                    child: PhotoDescriptionWidget(photoCard: photoCard),
+                  ),
                   SizedBox(
                     height: 20.0,
                   ),